Add files via upload
In this Sentiment Analysis project, I extensively explore the power to NLP libraries. In addition, data manipulation and feature engineering was the fun part. Data training and testing are crucial parts because the next steps are based on it. So, data cleaning, and normalization are very crucial aspects for machine learning.