Add files via upload
In this project, statistics for machine learning have been discussed comprehensively. This project extensively discusses descriptive statistics. In addition, density plots have been drawn to check the skewness of the dataset. Moreover, boxplots are also drawn to show outliers. The logarithmic Transformation technique is used to make data symmetrically distributed. Also, outliers are removed to avoid bias. In the end, data is normalized via standardization and a min-max scaler for Machine Learning.