Add files via upload
This is a very extensive project. In the beginning, data pre-processing is conducted to make data accurate for EDA and machine learning. Moreover, feature engineering helped to get hidden insights. Before, ML outliers are removed to avoid bias. In addition, standardization and a min-max scaler are used to distribute data normally. In the end, three algorithms are applied including; Support Vector Machine, Decision Tree, and K-Nearest Neighbours. These three algorithms are applied to one data to check the comparative results. In a nutshell, the Decision Tree algorithm was found to be accurate for this dataset. It showed a better accuracy rate than other algorithms